More MinGW build fixes

+++ src/providers/wms/qgstilescalewidget.cpp
@@ -117,7 +117,7 @@ void QgsTileScaleWidget::showTileScale( QMainWindow *mainWindow )
   }
 
   QgsMapCanvas *canvas = mainWindow->findChild<QgsMapCanvas *>( "theMapCanvas" );
-  QgsDebugMsg( QString( "canvas:%1 [%2]" ).arg(( ulong ) canvas, 0, 16 ).arg( canvas ? canvas->objectName() : "" ) );
+  QgsDebugMsg( QString( "canvas:%1 [%2]" ).arg(( intptr_t ) canvas, 0, 16 ).arg( canvas ? canvas->objectName() : "" ) );
   if ( !canvas )
   {
     QgsDebugMsg( "map canvas theMapCanvas not found" );
